Added Pre-Annotation Lambda Function and Tests
The lambda function serves the data to the AWS Groudtruth
in a form that it expects. This was added so that named variables
can be loaded from manifest into HTML for user display. The tests
involved changing from dashes to underscores for the package name.
This was done to import module succesfully.