Added ROIs class and util modules
Added ROIs class to hold required loading and morphological functions.
This was added to correctly identify the edge pixels of the inner
outline of an ROI in a 2D image. This was undertaken by identifying the
edges and the center then moving in along line by stroke size provided.
Query utils were added to load ROI data straight from LIMs db.