Changes to how videos are transformed
The mp4 videos were showing up green and broken on Mac devices because of an unsupported codec and output pixel type.  Movies are now output in webm format using the yuv420p pixel format. This change moves to the more widely supported codec and video container formats for cross platform usage. The tests were also changed to support this change to webm format. Additionally a small bug fix is introduced that fixes a problem with ffmpeg expecting size in opposite format as was being provided.