Apply image adjustment controls to projections
This also required some performance improvements.
Previously the canvases were triggered to redraw based
on slider input events, and many were fired as the sliders
were moving. This updates to redraw the canvas on a change
event, which is triggered when a final selection is made
on the slider.

Debugging the performance also brought to light that the
animation frame requests were happening too often, and not
stopping when the video stopped playing. The canvas drawing
behavior was moved to a separate function so that it could be
called (e.g. when brightness/contrast udpated) without starting animation.

Cancel animation when video is paused.