Adds a scale to the bottom of images and videos
This commit adds a scale onto the bottom of images and videos so that
users have a better idea of the size of objects in view for more informed
labeling. Added a new module, image_utils.py, to hold the add_scale function.
Add scale_function adds a scale of user specified size in microns
to the bottom of a supplied image. Transformation pipeline was updated
to make use of this function to add scales to outline masks, in full and
sub-sampled views. The testing functions were updated to work with this
change.